move sse code to github-events
