+++ lib/github-events.js
@@ -1,5 +1,5 @@
 const crypto = require('crypto')
-const debug = require('debug')('github')
+const debug = require('debug')('github-events')
 
 const secret = process.env.GITHUB_WEBHOOK_SECRET || 'hush-hush'
 
@@ -28,13 +28,31 @@ module.exports = (app) => {
     const data = req.body
     data.action = data.action ? event + '.' + data.action : event
 
-    var source = data.repository ? data.repository.full_name : data.organization.login
-    console.log('event@%s: %s', source, data.action)
-
-    if (debug.enabled) {
-      debug(JSON.stringify(data, null, 2))
-    }
-
-    app.emit(data.action, data)
+    emit(data)
   })
+
+  if (process.env.SSE_RELAY) {
+    const EventSource = require('eventsource')
+    var es = new EventSource(process.env.SSE_RELAY)
+    es.onmessage = (e) => {
+      try {
+        const data = JSON.parse(e.data)
+        if (!data.action) return
+
+        emit(data)
+      } catch (e) {
+        console.error(e)
+      }
+    }
+  }
+
+  function emit (data) {
+    const repo = data.repository
+    const org = repo.owner.login || data.organization.login
+
+    console.log('event@%s/%s: %s', org, repo.name, data.action)
+    debug(data)
+
+    app.emit(data.action, data, org, repo.name, data.sender.login)
+  }
 }
