+++ server.js
@@ -21,21 +21,3 @@ const port = process.env.PORT || 3000
 app.listen(port, () => {
   console.log('Example app listening on port', port)
 })
-
-if (process.env.SSE_RELAY) {
-  const EventSource = require('eventsource')
-  var es = new EventSource(process.env.SSE_RELAY)
-  es.onmessage = (e) => {
-    try {
-      const data = JSON.parse(e.data)
-      if (!data.action) return
-
-      var source = data.repository ? data.repository.full_name : data.organization.login
-      console.log('event@%s: %s', source, data.action)
-
-      app.emit(data.action, data)
-    } catch (e) {
-      console.error(e)
-    }
-  }
-}
